// src/api/models/PlayerTaskStatus.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum PlayerTaskStatus {
    PREPARING = 'PREPARING',
    IN_PROGRESS = 'IN_PROGRESS',
    PENDING_COMPLETION = 'PENDING_COMPLETION',
    COMPLETED = 'COMPLETED',
    SKIPPED = 'SKIPPED',
}